# second_service/src/main/java/com/example/model/Location.java
package com.example.model;

import lombok.Getter;
import lombok.Setter;
import lombok.NoArgsConstructor;

@Getter
@Setter
@NoArgsConstructor
public class Location {
    private Float x;
    private Float y;
    private String name;
} 